fix: rename routes to contain coumpoinds
