+++ node-pubchem/src/.babelrc
@@ -1,0 +1,3 @@
+{
+  "plugins": ["@babel/plugin-transform-modules-commonjs"]
+}
