+++ node-pubchem/src/package.json
@@ -47,6 +47,7 @@
     "workerpool": "^6.1.5"
   },
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.15.4",
     "@types/jest": "^27.0.2",
     "cheminfo-tools": "^1.23.3",
     "cross-env": "^7.0.3",
