+++ node-pubchem/src/src/server/v1/index.js
@@ -1,9 +1,9 @@
 import fastifySwagger from 'fastify-swagger';
 
 import { mfsFromEM } from './routes/compounds/mfsFromEM.js';
-import { moleculesFromEM } from './routes/compounds/moleculesFromEM.js';
-import { moleculesFromMF } from './routes/compounds/moleculesFromMF.js';
-import { moleculesFromSmiles } from './routes/compounds/moleculesFromSmiles.js';
+import { compoundsFromEM } from './routes/compounds/compoundsFromEM.js';
+import { compoundsFromMF } from './routes/compounds/compoundsFromMF.js';
+import { compoundsFromSmiles } from './routes/compounds/compoundsFromSmiles.js';
 
 export default function setupV1(app, _, done) {
   app.register(fastifySwagger, {
@@ -23,9 +23,9 @@ export default function setupV1(app, _, done) {
   });
 
   app.route(mfsFromEM);
-  app.route(moleculesFromEM);
-  app.route(moleculesFromMF);
-  app.route(moleculesFromSmiles);
+  app.route(compoundsFromEM);
+  app.route(compoundsFromMF);
+  app.route(compoundsFromSmiles);
 
   done();
 }
