--- node-pubchem/src/src/server/v1/routes/compounds/__tests__/moleculesFromEm.js
@@ -1,7 +1,0 @@
-import moleculesFromEM from '../../moleculesFromEM.js';
-
-test('moleculeByEm.js', async () => {
-  await expect(moleculesFromEM()).rejects.toStrictEqual(
-    new Error('em parameter must be specified'),
-  );
-});
