--- node-pubchem/src/src/server/v1/routes/compounds/__tests__/moleculesFromMf.js
@@ -1,7 +1,0 @@
-import moleculesFromEM from '../moleculesFromMf.js';
-
-test('moleculeByMf.js', async () => {
-  await expect(moleculesFromEM()).rejects.toStrictEqual(
-    new Error('mf parameter must be specified'),
-  );
-});
