+++ node-pubchem/src/src/sync/compound/utils/__tests__/improveCompound.test.js
@@ -1,7 +1,7 @@
 import { readFileSync } from 'fs';
 import { join } from 'path';
 
-import improveMolecule from '../improveMolecule.js';
+import improveCompound from '../improveCompound.js';
 
 test('creation of molecular formula', () => {
   let molfile = readFileSync(join(__dirname, 'test.mol'), 'utf8');
@@ -9,6 +9,6 @@ test('creation of molecular formula', () => {
     PUBCHEM_COMPOUND_CID: 1234,
     molfile,
   };
-  let result = improveMolecule(molecule);
+  let result = improveCompound(molecule);
   expect(result).toMatchSnapshot();
 });
