+++ node-pubchem/src/src/sync/compound/utils/__tests__/test.mol
