+++ node-pubchem/src/src/sync/substance/utils/__tests__/data.sdf
